# triedb/pathdb: improve the performance of parse index block