@rjl493456442 2025-07-16
IIUC, the performance speedup comes from the 1-loop traversal, instead of 2 right?